Provide a hash of the contract's variable state
This will improve efficiency because an independent script can call all of the pure/view functions, hash the result and then compare it with the functionDataId (hash) which is stored in Elasticsearch. If these two hash values are the same, the incremental updater will not even attempt to update the values stored in Elasticsearch.